# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Auracles API
  version: 0.1.0
  description: API contract for the Auracles knowledge marketplace.
servers:
  - url: http://localhost:8000
    description: Local FastAPI server
...
paths:
...
  /v1/admin/users/{user_id}/roles:
    patch:
      tags:
        - Admin
      operationId: assignUserRole
      summary: Assign or approve a user role
      description: Admin-only endpoint for assigning roles and approving Attestors.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRoleAssignmentRequest"
      responses:
        "200":
          description: Role assigned or approved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRoleAssignmentResponse"
        "401":
          description: Missing or invalid access token.
        "403":
          description: Admin role required.
        "404":
          description: User not found.
        "422":
          description: Invalid role payload.
  /v1/admin/users/{user_id}/kyc:
    patch:
      tags:
        - Admin
      operationId: reviewUserKyc
      summary: Review user KYC
      description: Admin-only endpoint for approving or rejecting a user's latest KYC document.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminKycReviewRequest"
      responses:
        "200":
          description: KYC review completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminKycReviewResponse"
        "401":
          description: Missing or invalid access token.
        "403":
          description: Admin role required.
        "404":
          description: User or KYC document not found.
        "422":
          description: Invalid review payload.
  /v1/admin/frameworks/{framework_id}/suspend:
    post:
      tags:
        - Admin
      operationId: suspendFramework
      summary: Suspend a published Framework
      description: Admin-only post-publish moderation endpoint that removes a Framework from marketplace discovery.
      security:
        - bearerAuth: []
      parameters:
        - name: framework_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFrameworkSuspendRequest"
      responses:
        "200":
          description: Framework suspended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFrameworkStatusResponse"
        "401":
          description: Missing or invalid access token.
        "403":
          description: Admin role required.
        "404":
          description: Framework not found.
        "409":
          description: Only published Frameworks can be suspended.
        "422":
          description: Invalid suspend payload.
  /v1/admin/licenses:
    post:
      tags:
        - Admin
      operationId: grantLicense
      summary: Grant a Framework license
      description: Admin-only Phase 2 endpoint for granting an Operator access to the current Framework version.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLicenseGrantRequest"
      responses:
        "201":
          description: License granted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLicenseGrantResponse"
        "401":
          description: Missing or invalid access token.
        "403":
          description: Admin role required.
        "404":
          description: Framework or Operator not found.
        "409":
          description: Operator already has a license for this Framework.
        "422":
          description: Invalid license grant payload.
...
  /v1/frameworks/{framework_id}/artifacts/{artifact_id}/download:
    get:
      tags:
        - Library
      operationId: downloadLicensedArtifact
      summary: Download licensed Artifact
      description: Returns a short-lived S3 GET URL after Operator role, verified KYC, active license, and version coverage checks pass.
      security:
        - bearerAuth: []
      parameters:
        - name: framework_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Licensed Artifact download URL returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactDownloadResponse"
        "401":
          description: Missing or invalid access token.
        "403":
          description: Operator role, verified KYC, active license, or licensed version required.
        "404":
          description: Artifact not found.
...
  /v1/library:
    get:
      tags:
        - Library
      operationId: listOperatorLibrary
      summary: List Operator library
      description: Returns Framework licenses held by the authenticated Operator.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Operator library returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryResponse"
        "401":
          description: Missing or invalid access token.
        "403":
          description: Operator role required.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminRoleAssignmentRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum:
            - contributor
            - operator
            - attestor
            - admin
    AdminRoleAssignmentResponse:
      type: object
      required:
        - user_id
        - role
        - approved
      properties:
        user_id:
          type: string
          format: uuid
        role:
          type: string
        approved:
          type: boolean
    AdminKycReviewRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - verified
            - rejected
        notes:
          type:
            - string
            - "null"
    AdminKycReviewResponse:
      type: object
      required:
        - user_id
        - kyc_status
        - document_status
      properties:
        user_id:
          type: string
          format: uuid
        kyc_status:
          type: string
        document_status:
          type: string
    AdminFrameworkSuspendRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
    AdminFrameworkStatusResponse:
      type: object
      required:
        - framework_id
        - status
      properties:
        framework_id:
          type: string
          format: uuid
        status:
          type: string
        reason:
          type:
            - string
            - "null"
    AdminLicenseGrantRequest:
      type: object
      required:
        - framework_id
        - operator_id
        - type
      properties:
        framework_id:
          type: string
          format: uuid
        operator_id:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - single_user
            - team
            - enterprise
        expires_at:
          type:
            - string
            - "null"
          format: date-time
        seats_total:
          type:
            - integer
            - "null"
          minimum: 1
    AdminLicenseGrantResponse:
      type: object
      required:
        - license_id
        - framework_id
        - operator_id
        - type
        - status
        - version_at_grant
        - seats_used
        - seats_total
        - expires_at
      properties:
        license_id:
          type: string
          format: uuid
        framework_id:
          type: string
          format: uuid
        operator_id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
        version_at_grant:
          type: string
        seats_used:
          type: integer
        seats_total:
          type:
            - integer
            - "null"
        expires_at:
          type:
            - string
            - "null"
          format: date-time
...
    LibraryItem:
      type: object
      required:
        - license_id
        - framework_id
        - title
        - version_at_grant
        - current_version
        - license_type
        - status
        - seats_used
        - seats_total
        - price
        - currency
        - thumbnail_key
        - granted_at
        - expires_at
      properties:
        license_id:
          type: string
          format: uuid
        framework_id:
          type: string
          format: uuid
        title:
          type: string
        version_at_grant:
          type: string
        current_version:
          type: string
        license_type:
          type: string
        status:
          type: string
        seats_used:
          type: integer
        seats_total:
          type:
            - integer
            - "null"
        price:
          type: string
          pattern: ^[0-9]+(\.[0-9]{1,2})?$
        currency:
          type: string
        thumbnail_key:
          type:
            - string
            - "null"
        granted_at:
          type: string
          format: date-time
        expires_at:
          type:
            - string
            - "null"
          format: date-time
    LibraryResponse:
      type: object
      required:
        - items
        - total
        - page
        - page_size
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryItem"
        total:
          type: integer
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
    ArtifactDownloadResponse:
      type: object
      required:
        - artifact_id
        - framework_id
        - license_id
        - download_url
        - expires_in
      properties:
        artifact_id:
          type: string
          format: uuid
        framework_id:
          type: string
          format: uuid
        license_id:
          type: string
          format: uuid
        download_url:
          type: string
        expires_in:
          type: integer
...